# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Provena",
    "summary": "Provenance-first memory service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/admin/retention-policies": {
      "post": {
        "summary": "Save Retention Policy",
        "operationId": "save_retention_policy_v1_admin_retention_policies_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionPolicy"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionPolicy"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Retention Policies",
        "operationId": "list_retention_policies_v1_admin_retention_policies_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RetentionPolicy"
                  },
                  "title": "Response List Retention Policies V1 Admin Retention Policies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/legal-hold": {
      "post": {
        "summary": "Place Legal Hold",
        "operationId": "place_legal_hold_v1_admin_legal_hold_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegalHold"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegalHold"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/legal-hold/{hold_id}": {
      "delete": {
        "summary": "Release Legal Hold",
        "operationId": "release_legal_hold_v1_admin_legal_hold__hold_id__delete",
        "parameters": [
          {
            "name": "hold_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hold Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Release Legal Hold V1 Admin Legal Hold  Hold Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/rtbf": {
      "post": {
        "summary": "Right To Be Forgotten",
        "operationId": "right_to_be_forgotten_v1_admin_rtbf_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RTBFRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RTBFResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/retention/enforce": {
      "post": {
        "summary": "Enforce Retention",
        "operationId": "enforce_retention_v1_admin_retention_enforce_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionEnforcementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionEnforcementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LegalHold": {
        "properties": {
          "hold_id": {
            "type": "string",
            "title": "Hold Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "memory_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Memory Ids"
          },
          "scope": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "hold_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hold Until"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "hold_id",
          "tenant_id",
          "reason"
        ],
        "title": "LegalHold"
      },
...
      "RTBFRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          }
        },
        "type": "object",
        "required": [
          "tenant_id"
        ],
        "title": "RTBFRequest"
      },
      "RTBFResponse": {
        "properties": {
          "deleted_memories": {
            "type": "integer",
            "title": "Deleted Memories"
          },
          "held_memories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Held Memories"
          }
        },
        "type": "object",
        "required": [
          "deleted_memories"
        ],
        "title": "RTBFResponse"
      },
...
      "RetentionEnforcementRequest": {
        "properties": {
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          }
        },
        "type": "object",
        "title": "RetentionEnforcementRequest"
      },
      "RetentionEnforcementResponse": {
        "properties": {
          "expired_memory_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Expired Memory Ids"
          }
        },
        "type": "object",
        "title": "RetentionEnforcementResponse"
      },
      "RetentionPolicy": {
        "properties": {
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "max_age_days": {
            "type": "integer",
            "title": "Max Age Days",
            "default": 365
          },
          "action": {
            "type": "string",
            "title": "Action",
            "default": "delete_soft"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "policy_id",
          "tenant_id"
        ],
        "title": "RetentionPolicy"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}